requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
get==0.0.39
idna==2.6
lxml==4.1.1
pickleDB==0.6.2
pkg-resources==0.0.0
post==0.0.26
public==0.0.65
python-magic==0.4.15
query-string==0.0.28
request==0.0.26
requests==2.18.4
simplejson==3.13.2
tinydb==3.7.0
urllib3==1.22
